\name{potlibR-package}
\alias{potlibR-package}
\alias{potlibR}
\docType{package}
\title{
\packageTitle{potlibR}
}
\description{
\packageDescription{potlibR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{potlibR}
\packageIndices{potlibR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{potlibR}

Maintainer: \packageMaintainer{potlibR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
